Extract postgres specific from vectorfilewriter test
